fix(feat-popup): 修改了addstyle和一些小问题
