+++ src/popup/popup.less
@@ -41,7 +41,11 @@
   text-overflow:ellipsis;
   overflow: hidden;
   color: #212121;
-  transition: color .3s transition-property color transition-duration 0.3s transition-timing-function ease transition-delay 0s;
+  transition: color .3s;
+  transition-property: color;
+  transition-duration: 0.3s;
+  transition-timing-function: ease;
+  transition-delay: 0s;
 }
 a.huiyou-title {
   height: 13px;
