+++ src/popup/showVideo.ts
@@ -1,8 +1,8 @@
-import axios from 'axios';
 import { PlayVideoInfo } from '@/types/video';
 import logger from '@/utils/logger';
 import { getVideoInfo } from '@/apis/bilibili';
-import { addClass } from './utils';
+import { CSSProperties } from 'vue';
+import { addClass, addStyle } from './utils';
 import { startTimekeeping } from './timeKeeper';
 
 let popupBox: HTMLDivElement | null = null;
@@ -14,7 +14,6 @@ let titleStr: Text | null = null;
 // let showVideo: VideoInfo | null = null;
 
 export function initialVideo() {
-  const { body } = document;
   titleBox = document.createElement('div');
   title = document.createElement('a');
   title.setAttribute('href', 'https://www.bilibili.com/video/');
@@ -24,7 +23,8 @@ export function initialVideo() {
   titleIcon.src = 'img/guanbi-1.svg';
   titleBox.appendChild(title);
   videoImg = document.createElement('div');
-  videoImg.style.backgroundImage = '';
+  addStyle(videoImg, 'backgroundImage: ' as CSSProperties);
+  // videoImg.style.backgroundImage = '';
   addClass(titleBox, 'huiyou-title-box');
   addClass(title, 'huiyou-title');
   addClass(videoImg, 'huiyou-pic');
@@ -41,31 +41,27 @@ export function initialBox() {
   const { body } = document;
   logger.info('init popup-box');
   popupBox = document.createElement('div');
-  popupBox.style.visibility = 'hidden';
+  addStyle(popupBox, 'visibility: hidden' as CSSProperties);
   body.appendChild(popupBox);
   addClass(popupBox, 'huiyou-popup-box');
   return popupBox;
 }
 export async function showVideo(bvid: string): Promise<PlayVideoInfo | null> {
-  let aid = 0;
-  let pic = '';
   let video: PlayVideoInfo | null = null;
   video = await getVideoInfo({ bvid }).then((data) => data.data.data);
   if (video !== null) {
-    aid = video.aid;
-    pic = video.pic;
     if (popupBox !== null && videoImg !== null && title !== null) {
       startTimekeeping(popupBox);
-      popupBox.style.visibility = 'visible';
+      addStyle(popupBox, 'visibility: visible' as CSSProperties);
       if (title.firstChild !== null) {
         title.removeChild(title.firstChild); // 删除原来的节点
       }
       titleStr = document.createTextNode(video.title);
       title.appendChild(titleStr);
       title.setAttribute('href', `https://www.bilibili.com/video/${video.bvid}`);
-      videoImg.style.backgroundImage = `url(${video.pic})`;
-      videoImg.style.backgroundRepeat = 'no-repeat';
-      videoImg.style.backgroundSize = '100% 100%';
+      addStyle(videoImg, `backgroundImage: url(${video.pic}` as CSSProperties);
+      addStyle(videoImg, 'backgroundRepeat: no-repeat' as CSSProperties);
+      addStyle(videoImg, 'backgroundSize: 100% 100%' as CSSProperties);
     }
   }
   if (video !== null) {
